Working version of the app with arrow key inputs.
* Add necessary functionality to `Model` in model and in `__init__` module to include arrow key press functionality.
* Remove forcing the `SpeedChanger` class to have to deal with `int` values; issues arose with actually getting the values to update.
* Move typing of scrolling movement speeds from `int` to `float` throughout codebase.
* Fix `test_model` tests that needed `SpeedChanger` patching.
* Add docstrings to classes in `speed_changer` module.

+++ tests/screen_animator/test_speed_changer.py
@@ -78,12 +78,12 @@ class TestSpeedChanger:
 
     @pytest.mark.parametrize(
         "speed_input, check_method",
-        [(Speed.FASTER, int.__gt__), (Speed.SLOWER, int.__lt__)],
+        [(Speed.FASTER, float.__gt__), (Speed.SLOWER, float.__lt__)],
     )
     def test_change_speed_changes(
         self,
         speed_input: Speed,
-        check_method: Callable[[int, int], bool],
+        check_method: Callable[[float, float], bool],
         example_speed_changer: SpeedChanger,
     ) -> None:
         """Speed of `Speeder` is changed."""
@@ -153,11 +153,11 @@ class TestMaintainSpeedAction:
 
 @pytest.mark.parametrize(
     "speed_action, check_method",
-    [(IncreaseSpeedAction, int.__gt__), (DecreaseSpeedAction, int.__lt__)],
+    [(IncreaseSpeedAction, float.__gt__), (DecreaseSpeedAction, float.__lt__)],
 )
 def test_speed_action_changes_speed(
     speed_action: type[SpeedAction],
-    check_method: Callable[[int, int], bool],
+    check_method: Callable[[float, float], bool],
     example_speed_changer: SpeedChanger,
 ) -> None:
     """Speed of `Speeder` is changed."""
